requirements.txt
# requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
tqdm==4.66.1